Authentification added
Added Login

Products can now be created, updated and deleted with authentification

+++ Website/public/publicScripts/getProduct.js
@@ -1,9 +1,20 @@
+// Whole Script: Copied and modified from https://github.com/BitSparkCode/online-shop-api
 document.addEventListener("DOMContentLoaded", () => {
 	const productNameText = document.getElementById("productName");
 	const productPriceText = document.getElementById("productPrice");
 	const productCategoryText = document.getElementById("productCategory");
 	const updateProductButton = document.getElementById("updateProductButton");
 	const deleteProductButton = document.getElementById("deleteProductButton");
+	const token = localStorage.getItem("token");
+	const feedbackText = document.getElementById("feedbackText");
+	const placeholder = document.getElementById("placeholder");
+	feedbackText.innerHTML = "";
+
+	if (!token) {
+		placeholder.remove();
+		deleteProductButton.style.display = "none";
+		updateProductButton.style.display = "none";
+	}
 
 	const fetchProduct = () => {
 		fetch(`/products/${window.location.pathname.split("/")[3]}`)
@@ -26,11 +37,24 @@ document.addEventListener("DOMContentLoaded", () => {
 	};
 
 	const deleteProduct = () => {
-		fetch(`/products/${window.location.pathname.split("/")[3]}`, {
-			method: "DELETE"
-		}).then(() => {
-			window.location.href = "/products";
-		});
+		if (token) {
+			fetch(`/products/${window.location.pathname.split("/")[3]}`, {
+				method: "DELETE",
+				// Authentification with token: https://reqbin.com/code/javascript/ricgaie0/javascript-fetch-bearer-token
+				headers: {
+					Authorization: `Bearer ${token}`
+				}
+			}).then(() => {
+				localStorage.setItem(
+					"productsFeedback",
+					"Produkt wurde gelöscht"
+				);
+				window.location.href = "/products";
+			});
+		} else {
+			feedbackText.innerHTML =
+				"Sie müssen sich authentifizieren, um Produkte zu löschen";
+		}
 	};
 
 	fetchProduct();
